+++ CHANGELOG.md
@@ -204,4 +204,12 @@ Chech for examples [GALLERY](https://github.com/NoHomey/react-material-ui-keyboa
 - Fixing possible styling bugs when other than `style` `prop`s, such as `inputStyle`, `underlineStyle` or any other posible custom `prop`, is passed to `textField` would make keyboard input "style boken". `textField` `style` `prop` `properties` `['minHeight', 'height', 'maxHeight']` are no longer `delete`d
 - Fixing possible styling bugs when `style` `prop` contains any of `['minHeight', 'height', 'maxHeight']` would prevent `Keyboard` from calculating it's height correctly
 
+## [v6.0.0](https://github.com/NoHomey/react-material-ui-keyboard/releases/tag/6.0.0)
+
+### Properties
+
+#### New
+
+- `correctionName` is a `string` which is the name of the cloned `textField` `prop` to which to bind `corrector`.
+- `corrector` is a `function`  which is bound to the the cloned `textField` at `correctorName` `prop`. `this` is bound to the `Keyboard`, `public` method `makeCorrection` can be used to apply a correction to the keyboard input.